Fetch managed domains from the API
