+++ src/components/clusterConfiguration/NetworkConfiguration.tsx
@@ -4,14 +4,18 @@ import { TextContent, Radio, Text, FormGroup } from '@patternfly/react-core';
 import BasicNetworkFields from './BasicNetworkFields';
 import AdvancedNetworkFields from './AdvancedNetworkFields';
 import { HostSubnets, ClusterConfigurationValues } from '../../types/clusters';
-import { InputField, CheckboxField } from '../ui/formik';
-import { RED_HAT_DNS_SERVICE_DOMAINS } from '../../config';
+import { InputField, CheckboxField, SelectField } from '../ui/formik';
+import { ManagedDomain } from '../../api/types';
 
 type NetworkConfigurationProps = {
   hostSubnets: HostSubnets;
+  managedDomains: ManagedDomain[];
 };
 
-const NetworkConfiguration: React.FC<NetworkConfigurationProps> = ({ hostSubnets }) => {
+const NetworkConfiguration: React.FC<NetworkConfigurationProps> = ({
+  hostSubnets,
+  managedDomains,
+}) => {
   const [type, setType] = React.useState<'basic' | 'advanced'>('basic');
   const { setFieldValue, initialValues, values } = useFormikContext<ClusterConfigurationValues>();
   const { name: clusterName, baseDnsDomain, useRedHatDnsService } = values;
@@ -35,33 +39,46 @@ const NetworkConfiguration: React.FC<NetworkConfigurationProps> = ({ hostSubnets
     </>
   );
 
-  const toggleRedHatDnsService = (checked: boolean) => {
-    if (checked) {
-      setFieldValue('baseDnsDomain', RED_HAT_DNS_SERVICE_DOMAINS[0]);
-    } else {
-      setFieldValue('baseDnsDomain', initialValues.baseDnsDomain);
-    }
-  };
+  const toggleRedHatDnsService = (checked: boolean) =>
+    setFieldValue(
+      'baseDnsDomain',
+      checked ? managedDomains.map((d) => d.domain)[0] : initialValues.baseDnsDomain,
+    );
 
   return (
     <>
       <TextContent>
         <Text component="h2">Networking</Text>
       </TextContent>
-      <CheckboxField
-        name="useRedHatDnsService"
-        label="Use Red Hat's DNS service"
-        helperText="A base DNS domain will be provided by Red Hat's DNS service. Because the cluster's DNS can't be changed later, this should only be used for temporary, non-production clusters."
-        onChange={toggleRedHatDnsService}
-      />
-      <InputField
-        label="Base DNS Domain"
-        name="baseDnsDomain"
-        helperText={baseDnsHelperText}
-        placeholder="example.com"
-        isDisabled={useRedHatDnsService}
-        isRequired
-      />
+      {!!managedDomains.length && (
+        <CheckboxField
+          name="useRedHatDnsService"
+          label="Use Red Hat's DNS service"
+          helperText="A base DNS domain will be provided by Red Hat's DNS service. Because the cluster's DNS can't be changed later, this should only be used for temporary, non-production clusters."
+          onChange={toggleRedHatDnsService}
+        />
+      )}
+      {values.useRedHatDnsService ? (
+        <SelectField
+          label="Base DNS Domain"
+          name="baseDnsDomain"
+          helperText={baseDnsHelperText}
+          options={managedDomains.map((d) => ({
+            label: `${d.domain} (${d.provider})`,
+            value: d.domain,
+          }))}
+          isRequired
+        />
+      ) : (
+        <InputField
+          label="Base DNS Domain"
+          name="baseDnsDomain"
+          helperText={baseDnsHelperText}
+          placeholder="example.com"
+          isDisabled={useRedHatDnsService}
+          isRequired
+        />
+      )}
       <FormGroup fieldId="networkConfigurationType" label="Network Configuration">
         <Radio
           id="networkConfigurationTypeBasic"
